move actions to `napari._app_model`
