+++ napari/_app_model/_app.py
@@ -8,6 +8,7 @@ from napari._app_model.actions._layerlist_context_actions import (
     LAYERLIST_CONTEXT_ACTIONS,
     LAYERLIST_CONTEXT_SUBMENUS,
 )
+from napari._app_model.actions._view import VIEW_ACTIONS
 
 APP_NAME = 'napari'
 
@@ -34,6 +35,7 @@ class NapariApplication(Application):
         self.injection_store.namespace = _napari_names  # type: ignore [assignment]
 
         self.register_actions(LAYERLIST_CONTEXT_ACTIONS)
+        self.register_actions(VIEW_ACTIONS)
         self.menus.append_menu_items(LAYERLIST_CONTEXT_SUBMENUS)
 
     @classmethod
