+++ napari/_qt/_qapp_model/qactions/_view.py
@@ -12,7 +12,6 @@ from app_model.types import (
 )
 
 from napari._app_model.constants import MenuGroup, MenuId
-from napari._qt._qapp_model.qactions._toggle_action import ViewerToggleAction
 from napari._qt.qt_main_window import Window
 from napari._qt.qt_viewer import QtViewer
 from napari.settings import get_settings
@@ -225,76 +224,3 @@ Q_VIEW_ACTIONS: list[Action] = [
         toggled=ToggleRule(get_current=_get_current_tooltip_visibility),
     ),
 ]
-
-MENUID_DICT = {'axes': MenuId.VIEW_AXES, 'scale_bar': MenuId.VIEW_SCALEBAR}
-
-toggle_action_details = [
-    (
-        'napari.window.view.toggle_viewer_axes',
-        trans._('Axes Visible'),
-        'axes',
-        'visible',
-    ),
-    (
-        'napari.window.view.toggle_viewer_axes_colored',
-        trans._('Axes Colored'),
-        'axes',
-        'colored',
-    ),
-    (
-        'napari.window.view.toggle_viewer_axes_labels',
-        trans._('Axes Labels'),
-        'axes',
-        'labels',
-    ),
-    (
-        'napari.window.view.toggle_viewer_axesdashed',
-        trans._('Axes Dashed'),
-        'axes',
-        'dashed',
-    ),
-    (
-        'napari.window.view.toggle_viewer_axes_arrows',
-        trans._('Axes Arrows'),
-        'axes',
-        'arrows',
-    ),
-    (
-        'napari.window.view.toggle_viewer_scale_bar',
-        trans._('Scale Bar Visible'),
-        'scale_bar',
-        'visible',
-    ),
-    (
-        'napari.window.view.toggle_viewer_scale_bar_box',
-        trans._('Scale Bar Box'),
-        'scale_bar',
-        'box',
-    ),
-    (
-        'napari.window.view.toggle_viewer_scale_bar_colored',
-        trans._('Scale Bar Colored'),
-        'scale_bar',
-        'colored',
-    ),
-    (
-        'napari.window.view.toggle_viewer_scale_bar_ticks',
-        trans._('Scale Bar Ticks'),
-        'scale_bar',
-        'ticks',
-    ),
-]
-
-# Add `Action`s that toggle various viewer `axes` and `scale_bar` sub-attributes
-# E.g., `toggle_viewer_scale_bar_ticks` toggles the sub-attribute `ticks` of the
-# viewer attribute `scale_bar`
-for cmd, cmd_title, viewer_attr, sub_attr in toggle_action_details:
-    Q_VIEW_ACTIONS.append(
-        ViewerToggleAction(
-            id=cmd,
-            title=cmd_title,
-            viewer_attribute=viewer_attr,
-            sub_attribute=sub_attr,
-            menus=[{'id': MENUID_DICT[viewer_attr]}],
-        )
-    )
